requirements.txt
Flask
yfinance
requests
python-dotenv
pandas
matplotlib

-e .